refactor: change slot data type

+++ src/ui/navigation.rs
@@ -1,7 +1,7 @@
-pub fn go_right(slot: u8) -> u8 {
+pub fn go_right(slot: usize) -> usize {
     if slot == 3 { 0 } else { slot + 1 }
 }
 
-pub fn go_left(slot: u8) -> u8 {
+pub fn go_left(slot: usize) -> usize {
     if slot == 0 { 3 } else { slot - 1 }
 }
